Move utils to their own module
bacon: test
Resolves: OKTA-89650
