+++ lib/index.js
@@ -10,13 +10,14 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-/* globals process, console, SDK_VERSION */
+/* globals SDK_VERSION */
 /*jshint maxcomplexity:15 */
 
 require('./vendor/polyfills');
 
 var Q                 = require('q');
 var util              = require('./util');
+var cookies           = require('./cookies');
 var sdkCrypto         = require('./crypto');
 var ajaxRequest       = require('ajaxRequest'); // request type determined at build time
 var OAuthError        = require('./errors/OAuthError');
@@ -24,161 +25,16 @@ var AuthApiError      = require('./errors/AuthApiError');
 var AuthSdkError      = require('./errors/AuthSdkError');
 var AuthPollStopError = require('./errors/AuthPollStopError');
 
-var LOG_PREFIX = '[OktaAuth]';
 var STATE_TOKEN_COOKIE_NAME = 'oktaStateToken';
 var FRAME_ID = 'okta-oauth-helper-frame';
 var DEFAULT_POLLING_DELAY = 500;
-var DEBUG = false;
-var IS_NODE = 'object' === typeof process && Object.prototype.toString.call(process) === '[object process]';
-var IS_BROWSER = !IS_NODE;
-
-// LOGGING
-function print(fn, args) {
-  if (DEBUG) {
-    var consoleArgs = Array.prototype.slice.call(args);
-    consoleArgs.unshift(LOG_PREFIX + ' ');
-    fn.apply(console, consoleArgs);
-  }
-}
-
-function log() {
-  if (typeof console !== 'undefined' && console.log) {
-    return print(console.log, arguments);
-  }
-}
-
-function error() {
-  if (typeof console !== 'undefined' && console.error) {
-    return print(console.error, arguments);
-  }
-}
-
-// UTILS
-function bind(fn, ctx) {
-  var additionalArgs = Array.prototype.slice.call(arguments, 2);
-  return function() {
-    var args = Array.prototype.slice.call(arguments);
-    args = additionalArgs.concat(args);
-    return fn.apply(ctx, args);
-  };
-}
-
-function isAbsoluteUrl(url) {
-  return /^(?:[a-z]+:)?\/\//i.test(url);
-}
-
-function isString(obj) {
-  return Object.prototype.toString.call(obj) === '[object String]';
-}
-
-function isObject(obj) {
-  return Object.prototype.toString.call(obj) === '[object Object]';
-}
-
-function isoToDate(str) {
-  var parts = str.match(/\d+/g),
-      isoTime = Date.UTC(parts[0], parts[1] - 1, parts[2], parts[3], parts[4], parts[5]),
-      isoDate = new Date(isoTime);
-
-  return isoDate;
-}
-
-function toQueryParams(obj) {
-  var str = [];
-  if (obj !== null) {
-    for (var key in obj) {
-      if (obj.hasOwnProperty(key) &&
-          obj[key] !== undefined &&
-          obj[key] !== null) {
-        str.push(key + '=' + encodeURIComponent(obj[key]));
-      }
-    }
-  }
-  if (str.length) {
-    return '?' + str.join('&');
-  } else {
-    return '';
-  }
-}
-
-function genRandomString(length) {
-  var randomCharset = 'abcdefghijklnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789';
-  var random = '';
-  for (var c = 0, cl = randomCharset.length; c < length; ++c) {
-    random += randomCharset[Math.floor(Math.random() * cl)];
-  }
-  return random;
-}
-
-function extend(obj1, obj2) {
-  for (var prop in obj2) {
-    if (obj2.hasOwnProperty(prop)) {
-      obj1[prop] = obj2[prop];
-    }
-  }
-}
-
-function removeNils(obj) {
-  var cleaned = {};
-  for (var prop in obj) {
-    if (obj.hasOwnProperty(prop)) {
-      var value = obj[prop];
-      if (value !== null && value !== undefined) {
-        cleaned[prop] = value;
-      }
-    }
-  }
-  return cleaned;
-}
-
-function clone(obj) {
-  if (obj) {
-    var str = JSON.stringify(obj);
-    if (str) {
-      return JSON.parse(str);
-    }
-  }
-  return obj;
-}
-
-// Analogous to _.omit
-function omit(obj) {
-  var props = Array.prototype.slice.call(arguments, 1);
-  var newobj = {};
-  for (var p in obj) {
-    if (obj.hasOwnProperty(p) && props.indexOf(p) == -1) {
-      newobj[p] = obj[p];
-    }
-  }
-  return clone(newobj);
-}
-
-function find(collection, searchParams) {
-  var c = collection.length;
-  while (c--) {
-    var item = collection[c];
-    var found = true;
-    for (var prop in searchParams) {
-      if (!searchParams.hasOwnProperty(prop)) {
-        continue;
-      }
-      if (item[prop] !== searchParams[prop]) {
-        found = false;
-        break;
-      }
-    }
-    if (found) {
-      return item;
-    }
-  }
-}
 
 function getLink(obj, linkName, altName) {
   if (!obj || !obj._links) {
     return;
   }
   
-  var link = clone(obj._links[linkName]);
+  var link = util.clone(obj._links[linkName]);
 
   // If a link has a name and we have an altName, return if they match
   if (link && link.name && altName) {
@@ -190,44 +46,8 @@ function getLink(obj, linkName, altName) {
   }
 }
 
-function setCookie(name, value, expiresAt) {
-  if (!IS_BROWSER) {
-    return;
-  }
-
-  var expiresText = '';
-  if (expiresAt) {
-    expiresText = ' expires=' + isoToDate(expiresAt).toUTCString() + ';';
-  }
-
-  var cookieText = name + '=' + value + ';' + expiresText;
-  log('Set cookie: ' + cookieText);
-  document.cookie = cookieText;
-
-  return cookieText;
-}
-
-function getCookie(name) {
-  if (!IS_BROWSER) {
-    return;
-  }
-
-  var pattern = new RegExp(name + '=([^;]*)'),
-    matched = document.cookie.match(pattern);
-
-  if (matched) {
-    var cookie = matched[1];
-    log('Got cookie: ', cookie);
-    return cookie;
-  }
-}
-
-function deleteCookie(name) {
-  setCookie(name, '', '1970-01-01T00:00:00Z');
-}
-
 function addStateToken(res, options) {
-  var builtArgs = clone(options) || {};
+  var builtArgs = util.clone(options) || {};
 
   // Add the stateToken if one isn't passed and we have one
   if (!builtArgs.stateToken && res.stateToken) {
@@ -244,7 +64,7 @@ function getStateToken(res) {
 function AuthTransaction(sdk, res) {
   if (res) {
     this.data = res;
-    extend(this, flattenEmbedded(sdk, res, res, {}));
+    util.extend(this, flattenEmbedded(sdk, res, res, {}));
     delete this.stateToken;
 
     // RECOVERY_CHALLENGE has some responses without _links.
@@ -340,7 +160,7 @@ function link2fn(sdk, res, obj, link, ref) {
         throw new AuthSdkError('Must provide a link name');
       }
 
-      var lk = find(link, {name: name});
+      var lk = util.find(link, {name: name});
       if (!lk) {
         throw new AuthSdkError('No link found for that name');
       }
@@ -369,7 +189,7 @@ function link2fn(sdk, res, obj, link, ref) {
 
           if (res.status === 'MFA_ENROLL') {
             // Add factorType and provider
-            extend(data, {
+            util.extend(data, {
               factorType: obj.factorType,
               provider: obj.provider
             });
@@ -380,14 +200,14 @@ function link2fn(sdk, res, obj, link, ref) {
             if (data.rememberDevice) {
               href += '?rememberDevice=true';
             }
-            data = omit(data, 'rememberDevice');
+            data = util.omit(data, 'rememberDevice');
 
           } else if (data.profile &&
                     data.profile.updatePhone !== undefined) {
             if (data.profile.updatePhone) {
               href += '?updatePhone=true';
             }
-            data.profile = omit(data.profile, 'updatePhone');
+            data.profile = util.omit(data.profile, 'updatePhone');
           }
 
           return postToTransaction(sdk, href, data);
@@ -433,7 +253,7 @@ function links2fns(sdk, res, obj, ref) {
 
 function flattenEmbedded(sdk, res, obj, ref) { /* jshint ignore:line */
   obj = obj || res;
-  obj = clone(obj);
+  obj = util.clone(obj);
 
   if (Array.isArray(obj)) {
     var objArr = [];
@@ -451,17 +271,17 @@ function flattenEmbedded(sdk, res, obj, ref) { /* jshint ignore:line */
     }
 
     // Flatten any nested _embedded objects
-    if (isObject(embedded[key]) || Array.isArray(embedded[key])) {
+    if (util.isObject(embedded[key]) || Array.isArray(embedded[key])) {
       embedded[key] = flattenEmbedded(sdk, res, embedded[key], ref);
     }
   }
 
   // Convert any links on the embedded object
   var fns = links2fns(sdk, res, obj, ref);
-  extend(embedded, fns);
+  util.extend(embedded, fns);
 
-  obj = omit(obj, '_embedded', '_links');
-  extend(obj, embedded);
+  obj = util.omit(obj, '_embedded', '_links');
+  util.extend(obj, embedded);
   return obj;
 }
 
@@ -473,40 +293,36 @@ function httpRequest(sdk, url, method, args, dontSaveResponse) {
     'Content-Type': 'application/json',
     'X-Okta-SDK': 'okta-auth-js-' + SDK_VERSION
   };
-  extend(headers, sdk.options.headers || {});
+  util.extend(headers, sdk.options.headers || {});
 
   var options = {
     headers: headers,
     data: args || undefined
   };
 
-  log('Request: ', method, url, options);
-
   var err, res;
   return new Q(sdk.options.ajaxRequest(method, url, options))
     .then(function(resp) {
-      log('Response: ', resp);
-
       res = resp.responseText;
-      if (res && isString(res)) {
+      if (res && util.isString(res)) {
         res = JSON.parse(res);
       }
 
       if (!dontSaveResponse) {
         if (!res.stateToken) {
-          deleteCookie(STATE_TOKEN_COOKIE_NAME);
+          cookies.deleteCookie(STATE_TOKEN_COOKIE_NAME);
         }
       }
 
       if (res && res.stateToken && res.expiresAt) {
-        setCookie(STATE_TOKEN_COOKIE_NAME, res.stateToken, res.expiresAt);
+        cookies.setCookie(STATE_TOKEN_COOKIE_NAME, res.stateToken, res.expiresAt);
       }
 
       return res;
     })
     .fail(function(resp) { 
       var serverErr = resp.responseText || {};
-      if (isString(serverErr)) {
+      if (util.isString(serverErr)) {
         try {
           serverErr = JSON.parse(serverErr);
         } catch (e) {
@@ -520,16 +336,14 @@ function httpRequest(sdk, url, method, args, dontSaveResponse) {
         serverErr.errorSummary = 'Unknown error';
       }
 
-      error('Error: ' + resp);
-
       if (sdk.options.transformErrorXHR) {
-        resp = sdk.options.transformErrorXHR(clone(resp));
+        resp = sdk.options.transformErrorXHR(util.clone(resp));
       }
 
       err = new AuthApiError(serverErr, resp);
 
       if (err.errorCode === 'E0000011') {
-        deleteCookie(STATE_TOKEN_COOKIE_NAME);
+        cookies.deleteCookie(STATE_TOKEN_COOKIE_NAME);
       }
 
       throw err;
@@ -537,12 +351,12 @@ function httpRequest(sdk, url, method, args, dontSaveResponse) {
 }
 
 function get(sdk, url, saveResponse) { /* jshint ignore:line */
-  url = isAbsoluteUrl(url) ? url : sdk.options.url + url;
+  url = util.isAbsoluteUrl(url) ? url : sdk.options.url + url;
   return httpRequest(sdk, url, 'GET', undefined, !saveResponse);
 }
 
 function post(sdk, url, args, dontSaveResponse) { /* jshint ignore:line */
-  url = isAbsoluteUrl(url) ? url : sdk.options.url + url;
+  url = util.isAbsoluteUrl(url) ? url : sdk.options.url + url;
   return httpRequest(sdk, url, 'POST', args, dontSaveResponse);
 }
 
@@ -563,8 +377,6 @@ function OktaAuth(args) {
       'Required usage: new OktaAuth({url: "https://sample.okta.com"})');
   }
 
-  DEBUG = args.debug;
-
   this.options = {
     url: args.url,
     clientId: args.clientId,
@@ -580,37 +392,35 @@ function OktaAuth(args) {
   }
 
   sdk.session = {
-    close: bind(closeSession, sdk, sdk),
-    exists: bind(sessionExists, sdk, sdk),
-    get: bind(getSession, sdk, sdk),
-    refresh: bind(refreshSession, sdk, sdk),
-    setCookieAndRedirect: bind(setCookieAndRedirect, sdk, sdk)
+    close: util.bind(closeSession, sdk, sdk),
+    exists: util.bind(sessionExists, sdk, sdk),
+    get: util.bind(getSession, sdk, sdk),
+    refresh: util.bind(refreshSession, sdk, sdk),
+    setCookieAndRedirect: util.bind(setCookieAndRedirect, sdk, sdk)
   };
 
   sdk.tx = {
-    status: bind(transactionStatus, sdk, sdk),
-    resume: bind(resumeTransaction, sdk, sdk),
-    exists: bind(transactionExists, sdk, sdk)
+    status: util.bind(transactionStatus, sdk, sdk),
+    resume: util.bind(resumeTransaction, sdk, sdk),
+    exists: util.bind(transactionExists, sdk, sdk)
   };
 
   // This is exposed so we can mock document.cookie in our tests
   sdk.tx.exists._getCookie = function(name) {
-    return getCookie(name);
+    return cookies.getCookie(name);
   };
 
   sdk.idToken = {
-    authorize: bind(getIdToken, sdk, sdk),
-    verify: bind(verifyIdToken, sdk, sdk),
-    refresh: bind(refreshIdToken, sdk, sdk),
-    decode: bind(decodeIdToken, sdk)
+    authorize: util.bind(getIdToken, sdk, sdk),
+    verify: util.bind(verifyIdToken, sdk, sdk),
+    refresh: util.bind(refreshIdToken, sdk, sdk),
+    decode: util.bind(decodeIdToken, sdk)
   };
 
   // This is exposed so we can mock window.location.href in our tests
   sdk.idToken.authorize._getLocationHref = function() {
     return window.location.href;
   };
-
-  log('OktaAuth created');
 }
 
 var proto = OktaAuth.prototype;
@@ -986,13 +796,13 @@ function getIdToken(sdk, oauthOptions, options) { /* jshint ignore:line */
     redirectUri: sdk.options.redirectUri || window.location.href,
     responseType: 'id_token',
     responseMode: 'okta_post_message',
-    state: genRandomString(64),
-    nonce: genRandomString(64),
+    state: util.genRandomString(64),
+    nonce: util.genRandomString(64),
     scope: ['openid', 'email']
   };
 
   // Add user-provided options
-  extend(oauthParams, oauthOptions);
+  util.extend(oauthParams, oauthOptions);
 
   // Start overriding any options that don't make sense
   var sessionTokenOverrides = {
@@ -1006,9 +816,9 @@ function getIdToken(sdk, oauthOptions, options) { /* jshint ignore:line */
   };
 
   if (oauthOptions.sessionToken) {
-    extend(oauthParams, sessionTokenOverrides);
+    util.extend(oauthParams, sessionTokenOverrides);
   } else if (oauthOptions.idp) {
-    extend(oauthParams, idpOverrides);
+    util.extend(oauthParams, idpOverrides);
   }
 
   // Quick validation
@@ -1017,7 +827,7 @@ function getIdToken(sdk, oauthOptions, options) { /* jshint ignore:line */
   }
 
   // Convert our params to their actual OAuth equivalents
-  var oauthQueryHash = removeNils({
+  var oauthQueryHash = util.removeNils({
     'client_id': oauthParams.clientId,
     'redirect_uri': oauthParams.redirectUri,
     'response_type': oauthParams.responseType,
@@ -1037,7 +847,7 @@ function getIdToken(sdk, oauthOptions, options) { /* jshint ignore:line */
   }
 
   // Use the query params to build the authorize url
-  var requestUrl = sdk.options.url + '/oauth2/v1/authorize' + toQueryParams(oauthQueryHash);
+  var requestUrl = sdk.options.url + '/oauth2/v1/authorize' + util.toQueryParams(oauthQueryHash);
 
   // Determine the flow type
   var flowType;
@@ -1169,7 +979,7 @@ function sessionExists(sdk) { /* jshint ignore:line */
 function getSession(sdk) { /* jshint ignore:line */
   return get(sdk, '/api/v1/sessions/me')
   .then(function(session) {
-    var res = omit(session, '_links');
+    var res = util.omit(session, '_links');
 
     res.refresh = function() {
       return post(sdk, getLink(session, 'refresh').href);
@@ -1198,7 +1008,7 @@ function refreshSession(sdk) { /* jshint ignore:line */
 function setCookieAndRedirect(sdk, sessionToken, redirectUrl) { /* jshint ignore:line */
   redirectUrl = redirectUrl || window.location.href;
   window.location = sdk.options.url + '/login/sessionCookieRedirect' +
-    toQueryParams({
+    util.toQueryParams({
       checkAccountSetupComplete: true,
       token: sessionToken,
       redirectUrl: redirectUrl
