+++ lib/reqwestRequest.js
@@ -20,7 +20,7 @@ function reqwestRequest(method, url, args) {
     data: JSON.stringify(args.data),
     withCredentials: true
   })
-  .then(function(resp) {
+  .then(function() {
     return r.request;
   });
   return r;
